trivkinsplus:kins module with extra joints 16jan19
  1) Support 'extra' joints that are homed using the
     existing homing state machine.  The 'extra' joints
     do not contribute to forward or inverse kinematics
     calculations.

  2) Create a hal component to manage independent motion
     planners (like limit3) for the 'extra' joints.

  3) includes: simulation configs demo1.ini, demo2.ini, an
     example M144 command and README, demo1.txt, demo2.txt
     files

  4) kins man page update

The implementation, based on trivkins.c, uses new files
extrajoints.[ch] to provide a setup routine
(extrajoints_setup()) with parameters for the number of
regular joints and the number of extra joints and a hal
function (extrajoints_update()).  The setup routine exports
the hal function and the hal pins needed to support an
external motion planner/controller both before and after
homing.

The trivkinsplus module accepts parameters coordinates=,
kinstype= and a new parameter: extrajoints=.

If the extrajoints= parameter is specified and nonzero:
  a) the coordinates= parameter *must* be set
  b) the kinematics type is set to KINEMATICS_BOTH
These provisions ensure that there are no numerical
gaps between the conventional kinematic joints and
the extra joints and ensures joint values are available
for display prior to homing in the axis gui.

The design method employed can be applied to other
kinematics modules by 1) including the file extrajoints.h,
2) invoking extrajoints_setup() with appropriate settings.

-----------------------------------------------------------------------
notes for force-updates
1) minor cleanup
2) add extrajoints.c and link module
3) fix Makefile: add extrajoints.o as prerequisite for trivkinsplus.o
4) demos use MIN,MAX, etc
5) behave like trivkins if extrajoints==0
6) manpage updates
7) fail on bad extrajoints=